[RESTEASY-1781] Solve issue with split package org.jboss.resteasy.api.validation
